feat: Fix the compilation path issue and standardize the code format
