+++ music_player2/docs/en/TECHNICAL.md
