--- music_player2/music_player2/audio_ctl.c
@@ -1,429 +1,0 @@
-/**
- * Audio Controller - Simulator Implementation
- * Provides audio playback simulation for development
- */
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <errno.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <fcntl.h>
-#include <string.h>
-#include <strings.h>
-#include <pthread.h>
-#include <sys/time.h>
-
-#define _GNU_SOURCE
-
-#include "audio_ctl.h"
-
-#define USING_SIMULATOR_AUDIO 1
-#ifndef AUDIO_DEBUG
-#define AUDIO_DEBUG 0
-#endif
-
-#ifdef AUDIO_DEBUG
-#include <syslog.h>
-#define AUDIO_LOG(fmt, ...) syslog(LOG_INFO, "[AUDIO] " fmt, ##__VA_ARGS__)
-#else
-#define AUDIO_LOG(fmt, ...)
-#endif
-
-// Functions
-static uint32_t estimate_mp3_duration(off_t file_size);
-static void* monitor_thread_func(void* arg);
-static void* playback_simulation_thread(void* arg) __attribute__((unused));
-
-/* Estimate MP3 duration based on file size */
-static uint32_t estimate_mp3_duration(off_t file_size)
-{
-    const uint32_t avg_bitrate = 128000; // Assume 128kbps average bitrate
-    const uint32_t bytes_per_sec = avg_bitrate / 8;
-    
-    if (bytes_per_sec > 0) {
-        return (uint32_t)(file_size / bytes_per_sec);
-    }
-    
-    return 240; // Default 4 minutes
-}
-
-/* Playback monitoring thread */
-static void* monitor_thread_func(void* arg)
-{
-    audioctl_s* ctl = (audioctl_s*)arg;
-    if (!ctl) {
-        AUDIO_LOG("Monitor thread: null parameter");
-        return NULL;
-    }
-    
-    AUDIO_LOG("Monitor thread started");
-    
-    while (ctl->monitor_running) {
-        usleep(100000); // Check every 100ms
-        
-        pthread_mutex_lock(&ctl->control_mutex);
-        
-        if (ctl->is_playing && !ctl->is_paused) {
-            ctl->current_position_ms += 100;
-            
-            if (ctl->current_position_ms >= ctl->total_duration_ms * 1000) {
-                AUDIO_LOG("Playback completed");
-                ctl->is_playing = false;
-                ctl->state = AUDIO_CTL_STATE_STOP;
-            }
-        }
-        
-        pthread_mutex_unlock(&ctl->control_mutex);
-    }
-    
-    AUDIO_LOG("Monitor thread exited");
-    return NULL;
-}
-
-/* Playback simulation thread */
-static void* playback_simulation_thread(void* arg)
-{
-    audioctl_s* ctl = (audioctl_s*)arg;
-    if (!ctl) {
-        return NULL;
-    }
-    
-    AUDIO_LOG("Simulation thread started: %s", ctl->file_path);
-    
-    while (ctl->is_playing && !ctl->should_stop) {
-        usleep(1000000); // Update every second
-        
-        pthread_mutex_lock(&ctl->control_mutex);
-        if (ctl->is_playing && !ctl->is_paused) {
-            // Simulate audio output processing
-        }
-        pthread_mutex_unlock(&ctl->control_mutex);
-    }
-    
-    AUDIO_LOG("Simulation thread ended");
-    return NULL;
-}
-
-/* Public API Functions */
-
-/* Detect audio file format by extension */
-int audio_ctl_detect_format(const char *path)
-{
-    if (!path) {
-        AUDIO_LOG("Null file path");
-        return AUDIO_FORMAT_UNKNOWN;
-    }
-    
-    size_t len = strlen(path);
-    if (len < 4) {
-        AUDIO_LOG("File path too short: %s", path);
-        return AUDIO_FORMAT_UNKNOWN;
-    }
-    
-    const char *ext = path + len - 4;
-    AUDIO_LOG("Detecting format: %s", ext);
-    
-    if (strcasecmp(ext, ".mp3") == 0) {
-        AUDIO_LOG("MP3 format detected: %s", path);
-        return AUDIO_FORMAT_MP3;
-    } else if (strcasecmp(ext, ".wav") == 0) {
-        AUDIO_LOG("WAV format detected: %s", path);
-        return AUDIO_FORMAT_WAV;
-    }
-    
-    AUDIO_LOG("Unknown audio format: %s", path);
-    return AUDIO_FORMAT_UNKNOWN;
-}
-
-/* Initialize audio controller */
-audioctl_s *audio_ctl_init_nxaudio(const char *path)
-{
-    if (!path) {
-        AUDIO_LOG("Init failed: null path");
-        return NULL;
-    }
-    
-    AUDIO_LOG("Initializing audio controller: %s", path);
-    
-    audioctl_s *ctl = (audioctl_s*)calloc(1, sizeof(audioctl_s));
-    if (!ctl) {
-        return NULL;
-    }
-    
-    strncpy(ctl->file_path, path, sizeof(ctl->file_path) - 1);
-    ctl->file_path[sizeof(ctl->file_path) - 1] = '\0';
-    
-    ctl->audio_format = audio_ctl_detect_format(path);
-    if (ctl->audio_format == AUDIO_FORMAT_UNKNOWN) {
-        AUDIO_LOG("Unsupported audio format");
-        free(ctl);
-        return NULL;
-    }
-    
-    struct stat st;
-    if (stat(path, &st) == 0) {
-        ctl->file_size = st.st_size;
-        AUDIO_LOG("File size: %lld bytes", (long long)ctl->file_size);
-        
-        if (ctl->audio_format == AUDIO_FORMAT_MP3) {
-            uint32_t duration_sec = estimate_mp3_duration(ctl->file_size);
-            ctl->total_duration_ms = duration_sec * 1000;
-            AUDIO_LOG("Estimated MP3 duration: %lu seconds", (unsigned long)duration_sec);
-        } else {
-            ctl->total_duration_ms = 240 * 1000; // Default 4 minutes
-        }
-    } else {
-        AUDIO_LOG("Cannot get file info: %s", strerror(errno));
-        free(ctl);
-        return NULL;
-    }
-    
-    ctl->nxplayer = (struct nxplayer_s*)0x12345678; // Simulator mock pointer
-    AUDIO_LOG("Created virtual NxPlayer instance: %p", ctl->nxplayer);
-    
-    if (pthread_mutex_init(&ctl->control_mutex, NULL) != 0) {
-        AUDIO_LOG("Failed to init mutex");
-        free(ctl);
-        return NULL;
-    }
-
-    // Initialize state
-    ctl->state = AUDIO_CTL_STATE_STOP;
-    ctl->is_playing = false;
-    ctl->is_paused = false;
-    ctl->should_stop = false;
-    ctl->current_position_ms = 0;
-    ctl->monitor_running = 0;
-    
-    AUDIO_LOG("Audio controller initialized");
-    return ctl;
-}
-
-/* Start audio playback */
-int audio_ctl_start(audioctl_s *ctl)
-{
-    if (!ctl) {
-        AUDIO_LOG("Start failed: null controller");
-        return -1;
-    }
-    
-    if (!ctl->nxplayer) {
-        AUDIO_LOG("Start failed: null NxPlayer");
-        return -1;
-    }
-
-    AUDIO_LOG("Starting simulation: %s", ctl->file_path);
-    AUDIO_LOG("Audio format: %d", ctl->audio_format);
-    
-    pthread_mutex_lock(&ctl->control_mutex);
-    
-    if (ctl->is_playing) {
-        AUDIO_LOG("Stopping current playback");
-        ctl->is_playing = false;
-        ctl->should_stop = true;
-    }
-    
-    if (access(ctl->file_path, R_OK) != 0) {
-        AUDIO_LOG("File access failed: %s, errno: %d (%s)", ctl->file_path, errno, strerror(errno));
-        pthread_mutex_unlock(&ctl->control_mutex);
-        return -1;
-    }
-    
-    AUDIO_LOG("File access OK, starting simulation...");
-    
-    // Simulate successful playback
-    ctl->state = AUDIO_CTL_STATE_START;
-    ctl->is_playing = true;
-    ctl->is_paused = false;
-    ctl->should_stop = false;
-    ctl->current_position_ms = 0;
-    
-    // Start monitoring thread
-    if (!ctl->monitor_running) {
-        ctl->monitor_running = 1;
-        if (pthread_create(&ctl->monitor_thread, NULL, monitor_thread_func, ctl) != 0) {
-            AUDIO_LOG("Failed to create monitoring thread");
-            // Continue, not a fatal error
-        } else {
-            AUDIO_LOG("Monitoring thread started");
-        }
-    }
-    
-    pthread_mutex_unlock(&ctl->control_mutex);
-    
-    AUDIO_LOG("Simulation playback started successfully");
-    return 0; // Success
-}
-
-// Pause playback
-int audio_ctl_pause(audioctl_s *ctl)
-{
-    if (!ctl || !ctl->nxplayer) {
-        AUDIO_LOG("Pause failed: controller or NxPlayer is null");
-        return -1;
-    }
-    
-    AUDIO_LOG("Pausing simulation playback");
-    
-    pthread_mutex_lock(&ctl->control_mutex);
-    
-    if (!ctl->is_playing || ctl->is_paused) {
-        AUDIO_LOG("Currently not playing or already paused");
-        pthread_mutex_unlock(&ctl->control_mutex);
-        return 0;
-    }
-    
-    ctl->state = AUDIO_CTL_STATE_PAUSE;
-    ctl->is_paused = true;
-    AUDIO_LOG("Simulation pause successful");
-    
-    pthread_mutex_unlock(&ctl->control_mutex);
-    return 0;
-}
-
-// Resume playback
-int audio_ctl_resume(audioctl_s *ctl)
-{
-    if (!ctl || !ctl->nxplayer) {
-        AUDIO_LOG("Resume playback failed: controller or NxPlayer is null");
-        return -1;
-    }
-    
-    AUDIO_LOG("Resuming simulation playback");
-    
-    pthread_mutex_lock(&ctl->control_mutex);
-    
-    if (!ctl->is_playing || !ctl->is_paused) {
-        AUDIO_LOG("Currently not paused");
-        pthread_mutex_unlock(&ctl->control_mutex);
-        return 0;
-    }
-    
-    ctl->state = AUDIO_CTL_STATE_START;
-    ctl->is_paused = false;
-    AUDIO_LOG("Resume simulation playback successful");
-    
-    pthread_mutex_unlock(&ctl->control_mutex);
-    return 0;
-}
-
-// Stop playback
-int audio_ctl_stop(audioctl_s *ctl)
-{
-    if (!ctl || !ctl->nxplayer) {
-        AUDIO_LOG("Stop playback failed: controller or NxPlayer is null");
-        return -1;
-    }
-    
-    AUDIO_LOG("Stopping simulation playback");
-    
-    pthread_mutex_lock(&ctl->control_mutex);
-    
-    // Stop monitoring thread
-    if (ctl->monitor_running) {
-        ctl->monitor_running = 0;
-        ctl->should_stop = true;
-        pthread_mutex_unlock(&ctl->control_mutex);
-        
-        // Wait for monitoring thread to end
-        if (ctl->monitor_thread != 0) {
-            pthread_join(ctl->monitor_thread, NULL);
-            ctl->monitor_thread = 0;
-            AUDIO_LOG("Monitoring thread stopped");
-        }
-        
-        pthread_mutex_lock(&ctl->control_mutex);
-    }
-    
-    ctl->state = AUDIO_CTL_STATE_STOP;
-    ctl->is_playing = false;
-    ctl->is_paused = false;
-    ctl->current_position_ms = 0;
-    AUDIO_LOG("Stop simulation playback successful");
-    
-    pthread_mutex_unlock(&ctl->control_mutex);
-    return 0;
-}
-
-// Set volume
-int audio_ctl_set_volume(audioctl_s *ctl, uint16_t vol)
-{
-    if (!ctl || !ctl->nxplayer) {
-        AUDIO_LOG("Set volume failed: controller or NxPlayer is null");
-        return -1;
-    }
-    
-    // Ensure volume is within valid range
-    if (vol > 100) {
-        vol = 100;
-    }
-    
-    AUDIO_LOG("Set simulation volume: %d", vol);
-    return 0;
-}
-
-// Get current playback position
-int audio_ctl_get_position(audioctl_s *ctl)
-{
-    if (!ctl) {
-        return 0;
-    }
-    
-    pthread_mutex_lock(&ctl->control_mutex);
-    uint32_t position_sec = ctl->current_position_ms / 1000;
-    pthread_mutex_unlock(&ctl->control_mutex);
-    
-    return (int)position_sec;
-}
-
-// Seek to specified position
-int audio_ctl_seek(audioctl_s *ctl, unsigned ms)
-{
-    if (!ctl) {
-        AUDIO_LOG("Seek failed: controller is null");
-        return -1;
-    }
-    
-    AUDIO_LOG("Simulation seek to position: %lu ms", (unsigned long)ms);
-    
-    pthread_mutex_lock(&ctl->control_mutex);
-    
-    if (ms <= ctl->total_duration_ms) {
-        ctl->current_position_ms = ms;
-        AUDIO_LOG("Simulation position update successful: %lu ms", (unsigned long)ms);
-    } else {
-        AUDIO_LOG("Seek position exceeds file length: %lu ms > %lu ms", (unsigned long)ms, (unsigned long)ctl->total_duration_ms);
-        pthread_mutex_unlock(&ctl->control_mutex);
-        return -1;
-    }
-    
-    pthread_mutex_unlock(&ctl->control_mutex);
-    return 0;
-}
-
-// Release audio controller
-int audio_ctl_uninit_nxaudio(audioctl_s *ctl)
-{
-    if (!ctl) {
-        return 0;
-    }
-    
-    AUDIO_LOG("Releasing simulator audio controller");
-    
-    // Stop playback
-    audio_ctl_stop(ctl);
-    
-    // No need to release real NxPlayer in simulator environment
-    ctl->nxplayer = NULL;
-    
-    // Destroy mutex
-    pthread_mutex_destroy(&ctl->control_mutex);
-    
-    // Free memory
-    free(ctl);
-    
-    AUDIO_LOG("Simulator audio controller release completed");
-    return 0;
-}
